+++ src/wget_progress.h
@@ -31,10 +31,6 @@
 #include "wget_job.h"
 #include "wget_main.h"
 
-#define DEFAULT_SCREEN_WIDTH 70
-
-#define MINIMUM_SCREEN_WIDTH 45
-
 // the following is just needed for the progress bar
 struct _body_callback_context {
 	JOB *job;
